Cleanup cred implementation
Cleans up the credential implementation to better handle the marshalling
API. This includes splitting out those APIs into a separate pyx file as
it was introduced in MIT krb5 1.20 and also unified how the krb5_creds
value is stored in the Python object that represents it.
